Revert "Add a Site class and rendering"
This reverts commit bb6a4ac8c0fe34b4c346566450087f06e9bf40b2.

I feel like adding the site and renderer early on like this leads me
down a path of building a static site generator, while I'd rather
build a set of tools that I can myself combine to build tiny
bespoke static site builders.

So, I'd like to keep it in the history, but I'd like to try
keeping `pygeon` simpler and once I've created a couple of
websites manually with it, only then create a small static
site generator to go with it, maybe as a separate repo.